src/requirements.txt
altair==4.2.2
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
entrypoints==0.4
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.34.1
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
six==1.17.0
smmap==5.0.2
streamlit==1.17.0
sympy==1.13.3
toml==0.10.2
toolz==1.0.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.108
ultralytics-thop==2.0.14
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.21.0
